general cleanup and schema updates

+++ magpie/cli/purge_expired_network_tokens.py
@@ -46,6 +46,12 @@ def main(args=None, parser=None, namespace=None):
     setup_logger_from_options(LOGGER, args)
     db_session = get_db_session_from_config_ini(args.ini_config)
     deleted = models.NetworkToken.get_expired(db_session).delete()
+    anonymous_network_user_ids = [n.anonymous_user().id for n in db_session.query(models.NetworkNode).all()]
+    # clean up unused records in the database (no need to keep records associated with anonymous network users)
+    (db_session.query(models.NetworkRemoteUser)
+     .filter(models.NetworkRemoteUser.user_id.in_(anonymous_network_user_ids))
+     .filter(models.NetworkRemoteUser.network_token_id == None)   # noqa: E711
+     .delete())
     try:
         transaction.commit()
         db_session.close()
